% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildingData.R
\docType{data}
\name{building}
\alias{building}
\title{The number of people entering and exiting the CalIt2 building.}
\format{A data frame with 5040 rows and 10 variables:
   \describe{
       \item{Day}{The weekday of the day of interest.}
       \item{Date}{The date of interest, YYYY-DD-MM.}
       \item{Time}{The time of interest, hh:mm:ss.}
       \item{FlowIn}{Number of people entering the building reported for the
          previous half hour.}
       \item{FlowOut}{Number of people leaving the building reported for the
          previous half hour.}
       \item{EventId}{Marks the occurrence of an event. Takes the value 0
          when no event occurs, and values 1, 2, or 3 when event is happening.
          The value is specific to the type of event (anonymous).}
       \item{EventStart}{An indicator variable marking that there is an event
          to start within the next hour.}
       \item{EventEnd}{An indicator variable marking that an event has ended
          within the last hour.}
       \item{Weekday}{An indicator variable signalling that the day in question
          is a weekday (and therefore a work day).}
       \item{DayTime}{An indicator variable signalling the 12 hours between
          07:30 and 19:30 (what we will call 'day time').}
   }}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/event-detection/}
}
\usage{
building
}
\description{
The number of people entering and exiting the CalIt2 building at
   UCI is recorded. Originally, this data was used for event detection analysis,
   but we model it as a bivariate time series of counts.
}
\examples{
data(building)

}
\references{
\insertRef{IhlerEtAl2006}{PoARX}
}
\keyword{datasets}
