% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaFunctions.R
\name{dFranksCopulaPoisson}
\alias{dFranksCopulaPoisson}
\title{Frank's copula PMF for variables from a Poisson model}
\usage{
dFranksCopulaPoisson(obs, means, rho, ...)
}
\arguments{
\item{obs}{a matrix of non-negative integers. Each row represents observed
values of a (vector) count variable, whose probability is to be computed,
see Details.}

\item{means}{a matrix of non-negative means. It should have the same
dimensions as \code{obs}.}

\item{rho}{A numeric value specifying the dependence parameter.}

\item{...}{currently not used.}
}
\value{
a numeric vector whose i-th element gives the joint probability of
    the values in the i-th row of \code{obs}.
}
\description{
Compute the probability mass function for a Frank's copula with
    Poisson marginals.
}
\details{
Let \eqn{m} be the number of columns of \code{obs}. Each row of
    \code{obs} is a vector of observations of a vector r.v. with Poisson
    marginals connected by a Frank's copula with parameter \code{rho}.  The
    means of the Poisson marginals are in the corresponding row of
    \code{means}.

    For each row \eqn{o} of \code{obs}, \code{dFranksCopulaPoisson} computes
    \eqn{P(Y_1 = o_1, \dots, Y_m = o_m )}, where \eqn{Y_j} are Poisson random
    variables with means specified by the corresponding row of \code{means}
    and dependence parameter rho.

    In this package each column of \code{obs} represents a count time series,
    so row \code{t} is a vector of observations at time \code{t}.  The means,
    \code{means}, are usually computed by \code{PoARXFilter()}.
}
\references{
\insertRef{Nelsen2006}{PoARX}
}
\seealso{
\code{\link{pFranksCopula}}, \code{\link{PoARXFilter}}
}
