% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXEstimation.R
\name{.PoARXScore}
\alias{.PoARXScore}
\title{Marginal PoARX score working function}
\usage{
.PoARXScore(theta, obs, exogenous = NULL, intercept = TRUE,
  ylags = NULL, mulags = NULL, yinit = NULL, muinit = NULL, ...)
}
\arguments{
\item{theta}{A vector of parameter values.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{exogenous}{A 3-dimensional array containing the values of the exogenous
covariates. Each new value in dimension 1 represents another observation
of the time series. Each new vector in dimension 2 corresponds to a
new covariate and each matrix of the third dimension represents a
separate time series.}

\item{intercept}{A logical indicator for whether an intercept should be
included in the model.}

\item{ylags}{A vector of time lags in the filter equation relating to the
observations.}

\item{mulags}{A vector of time lags in the filter equation relating to the
intensities.}

\item{yinit}{A matrix detailing the initialisations of the observations.}

\item{muinit}{A matrix detailing the initialisations of the means.}

\item{...}{currently not used.}
}
\value{
An array of dimension \code{(n, N, K)} containing the score vector for
   each observation given the parameters.
}
\description{
Find the score function for observations from the marginal distributions of
   PoARX models. This returns the score for each individual observation.
}
\details{
\code{n} is the number of observations, \code{N} is the dimension of the
   score vector and \code{K} is the number of time series.
}
\keyword{internal}
