% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXPredict.R
\name{.predictLambda}
\alias{.predictLambda}
\title{Predict wrapper for the intensity values of a PoARX model}
\usage{
.predictLambda(object, newdata)
}
\arguments{
\item{object}{A \code{"PoARX"} object from \code{PoARXFit}.}

\item{newdata}{An optional data.frame for the new covariates and observations.
The names of the data.frame should match the original variable names. If some new
observations have not been obtained, put an NA value where the missing value is
and the observation will be estimated by \code{PoARXPredict}.}

\item{type}{A character argument indicating the type of response. If \code{"response"},
return the mean parameter for each time series. If \code{"probability"}, give the
probability of the response.}

\item{interval}{A logical argument for the calculation of confidence intervals.}

\item{cl}{A numeric value representing the desired confidence level for predictions of
response values.}

\item{B}{A number of desired simulations in the calculation of confidence intervals in
any \code{'response'} values.}

\item{trace}{A logical indicator for the printing of update messages.}

\item{...}{currently not used.}
}
\value{
If \code{type = "response"} a data.frame is returned containing
   \eqn{\lambda} values for each time series, along with optional confidence intervals.
   If \code{type = "probability"} and the respective observations are provided, then
   a matrix of probabilities is returned. If no observations are given, there
   are too many possible values, so the \eqn{\lambda} values are returned instead,
   as if \code{type = "response"}.
}
\description{
Used in the \code{predict.PoARX} function and contains the
   "hard work" of the predictions. New data can be fed into the model and
   a prediction given - this can be done h steps into the future, where
   the variance will grow with h. Missing observation values are estimated by
   the respective intensity values. If a response is to be returned, there is an
   option to form a confidence interval for this value, using simulation.
}
\keyword{internal}
