% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaFunctions.R
\name{.rhoFCInfoMat}
\alias{.rhoFCInfoMat}
\title{The profile information matrix for the dependence parameter rho.}
\usage{
.rhoFCInfoMat(rho, obs, means, zeroFix = 1e-35)
}
\arguments{
\item{rho}{A numeric value specifying the dependence parameter used in
\code{pFranksCopula}.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{means}{A matrix of calculated means, from \code{PoARXFilter}. It should
have the same dimensions as \code{obs}.}

\item{zeroFix}{A numeric value to transform any 0-probability values to.}
}
\value{
A numeric value representing the log-likelihood of the data for
   dependence parameter rho.
}
\description{
The negative value from twice differentiating the log-likelihood
    function for a model coupling Poisson distributions with a Frank's copula
    with respect to the dependence parameter.
}
\keyword{internal}
