% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaFunctions.R
\name{.rhothetaFCInfoMat}
\alias{.rhothetaFCInfoMat}
\title{Information matrix for the rho-theta part of PoARX model}
\usage{
.rhothetaFCInfoMat(vartheta, obs, means, xlist = NULL, modelSummary,
  ylagList = NULL, mulagList = NULL, yinitList = NULL,
  muinitList = NULL, sameParams = FALSE, zeroFix = 1e-35, ...)
}
\arguments{
\item{vartheta}{A vector of parameter values - all marginal parameters and the
dependence parameter.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{means}{A matrix of calculated means, from \code{PoARXFilter}. It should
have the same dimensions as \code{obs}.}

\item{xlist}{A list containing matrices of exogenous covariates. Each row
represents a new observation, column represents a new covariate. Each element
of the list is for each time series.}

\item{modelSummary}{A data.frame summarising the features of the PoARX models.}

\item{ylagList}{A list containing vectors of lags for observations. Each component
of the list represents a time series.}

\item{mulagList}{A list containing vectors of lags for intensities. Each component
of the list represents a time series.}

\item{yinitList}{A list of matrices detailing the initialisations of the observations.}

\item{muinitList}{A list of matrices detailing the initialisations of the means.}

\item{sameParams}{A logical argument indicating whether all marginal
distributions have the same parameter values.}

\item{zeroFix}{A numeric value to transform any 0-probability values to.}

\item{...}{currently not used.}
}
\value{
A numeric vector representing the score function of the observations
   given the parameters.
}
\description{
Find the information matrix for the theta derivative of the rho derivative
   of the PoARX likelihood.
}
\keyword{internal}
