% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitPoARX}
\alias{fitPoARX}
\title{Fit a PoARX model}
\usage{
fitPoARX(formula, data, subset, init, ylags = NULL, mulags = NULL,
  yinit = NULL, muinit = NULL, indep = FALSE, sameParams = FALSE,
  maxRho = 1000, zeroFix = 1e-50, ...)
}
\arguments{
\item{formula}{A formula object representing the response values and exogenous
covariates. Different time series can be separated using \code{|}.}

\item{data, subset}{arguments controlling formula processing via \code{model.frame}.}

\item{init}{An initial vector of parameter values.}

\item{ylags}{A vector or matrix of time lags in the filter equation relating to the
observations. If a matrix, the number of columns should correspond to the number
of time series and each column gives the lags of the corresponding time series.
When the lengths of lags do not match, use zeroes to fill the matrix.}

\item{mulags}{A vector or matrix of time lags in the filter equation relating to the
intensities. If a matrix, the number of columns should correspond to the number
of time series and each column gives the lags of the corresponding time series.
When the lengths of lags do not match, use zeroes to fill the matrix.}

\item{yinit}{A matrix detailing the initialisations of the observations. If the
maximum lags do not match, fill the required entries with zeroes.}

\item{muinit}{A matrix detailing the initialisations of the means. If the
maximum lags do not match, fill the required entries with zeroes.}

\item{indep}{A logical argument indicating whether the PoARX models are
coupled using Frank's copula or the independence copula.}

\item{sameParams}{A logical argument indicating whether the time series
should be modelled using the same set of PoARX parameter values.}

\item{maxRho}{A value indicating the largest absolute value that the dependence
parameter of Frank's copula can take. For use in \code{optimise}.}

\item{zeroFix}{A numeric value to transform any 0-probability values to.}

\item{...}{currently not used.}
}
\value{
An object of class 'PoARX'. The structure of the object is internal
    and subject to change. Currently it contains the following components:

    \item{formula}{model formula, an object from class \code{"Formula".}}
    \item{data}{the model frame (?)}
    \item{X}{}
    \item{Y}{}
    \item{nobs, nts}{The number of observations and time series, respectively.}
    \item{npar}{}
    \item{modelSummary}{}
    \item{ylags}{}
    \item{ylagList}{}
    \item{mulags}{}
    \item{mulagList}{}
    \item{sameParams}{}
    \item{rhoInd}{Information regarding the format of the PoARX model.}
    \item{yinit, muinit}{Information regarding the format of the PoARX model.}
    \item{rhoValue}{}
    \item{theta}{The estimated parameters.}
    \item{logLik}{The log-likelihood value for the parameters given the data.}
    \item{varCovar}{The variance-covariance matrix of the parameters.}
    \item{fitted.values}{}
    \item{fitted.probs}{}
    \item{intercept}{\strong{TODO: is this really present?}}
}
\description{
Fit a PoARX models to \code{K} time series using the method of
    inference functions \insertCite{Joe2005}{PoARX}.
}
\details{
A time series, \eqn{Y_t}, following a PoARX model obeys the following equations:
      \deqn{Y_t = N_t(\lambda_t),}
      \deqn{\lambda_t = \omega + \sum \alpha_l Y_{t-j_l} + \sum \beta_l
               \lambda_{t-j_l} + \eta x_{t-1},}
   where \eqn{N_t(.)} represents a Poisson process of unit intensity, so that
   \eqn{N_t(\lambda_t)} represents a Poisson process of intensity \eqn{\lambda_t}.
   \eqn{Y_t} represents the observations and \eqn{x_t} the covariates at time \eqn{t}.
   \eqn{\omega}, {\eqn{\alpha_i}}, {\eqn{\beta_i}} and \eqn{\eta} are all positive
   parameters that must be supplied. An additional constraint on the parameters is:
      \deqn{\sum (\alpha_i + \beta_i) < 1,}
   which means that the model satisfies the conditions for stationarity and
   ergodicity. Multiple time series are modelled independently by this fit
   procedure. The number of time series is denoted by \code{K}, the number of past
   observations by \code{p}, the number of past means by \code{q}, the number of
   exogenous covariates by \code{r}, and \code{w} is an indicator of an intercept
   in the model.
}
\examples{
#### examples here!

}
\references{
\insertAllCited{}
}
