% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXEstimation.R
\name{fullPoARXLogLikelihood}
\alias{fullPoARXLogLikelihood}
\title{Full PoARX log-likelihood}
\usage{
fullPoARXLogLikelihood(vartheta, obs, xlist, modelSummary, ylagList,
  mulagList, yinitList, muinitList, indep = FALSE, sameParams = FALSE,
  zeroFix = 1e-35, ...)
}
\arguments{
\item{vartheta}{A vector of parameter values, including all marginal parameters
and any dependence parameters.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{xlist}{A list of matrices containing the values of the exogenous
covariates. Each new value in dimension 1 represents another observation
of the time series. Each new vector in dimension 2 corresponds to a
separate new covariate. Each component of the list represents a
time series.}

\item{modelSummary}{A data.frame summarising the features of the PoARX models.}

\item{ylagList}{A list of vectors of time lags in the filter equation
relating to the observations.}

\item{mulagList}{A list of vectors of time lags in the filter equation
relating to the intensities.}

\item{yinitList}{A list of matrices detailing the initialisations of the
observations.}

\item{muinitList}{A list of matrices detailing the initialisations of the means.}

\item{indep}{A logical argument indicating whether the PoARX models are
coupled using Frank's copula or the independence copula.}

\item{sameParams}{A logical argument indicating whether the time series
should be modelled using the same set of PoARX parameter values.}

\item{zeroFix}{A numeric value to transform any 0-probability values to.}

\item{...}{currently not used.}
}
\value{
A numeric value representing the log-likelihood of the observations
   given the parameters.
}
\description{
Find the log-likelihood for observations from a PoARX model.
}
