% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARX-package.R
\docType{package}
\name{PoARX-package}
\alias{PoARX-package}
\alias{poarx}
\alias{PoARX}
\title{Fit PoARX models to multivariate time series}
\description{
PoARX models for multivariate time series. The time series can
   be independent, or modelled with a dependency using Frank’s Copula.
}
\details{
We fit PoARX models to multivariate time series count data using
   \code{PoARXFit}. The univariate PoARX model was
   first introduced as a method of modelling corporate defaults by Agosto
   et al. (2016). A time series of counts, \eqn{Y_t},
   following a univariate PoARX model obeys the following equations:
      \deqn{Y_t = N_t(\lambda_t),}
      \deqn{\lambda_t = \omega + \sum \alpha_l Y_{t-j_l} + \sum \beta_l
               \lambda_{t-j_l} + \eta x_{t-1},}
   where \eqn{N_t(.)} represents a Poisson process of unit intensity, so that
   \eqn{N_t(\lambda_t)} represents a Poisson process of intensity \eqn{\lambda_t}.
   \eqn{Y_t} represents the observations and \eqn{x_t} the covariates at time \eqn{t}.
   \eqn{\omega}, {\eqn{\alpha_i}}, {\eqn{\beta_i}} and \eqn{\eta} are all positive
   parameters that must be supplied. An additional constraint on the parameters is:
      \deqn{\sum (\alpha_i + \beta_i) < 1,}
   which means that the model satisfies the conditions for stationarity and
   ergodicity.

   With our package, multivariate time series can be modelled independently or coupled
   using Franks copula, accoring to the logical argument \code{indep} in \code{PoARXFit}.
   For the copula model, we use a generator and its pseudo-inverse to produce the
   required dependence structure. The generator for Frank's copula is given by
      \deqn{\varphi_\rho(t) = -\log ( (\exp(- \rho t ) - 1 ) /
         (\exp(- \rho) - 1 ) ),}
   and the pseudo-inverse by
      \deqn{\varphi^{[-1]}_\rho(t) = - 1/\rho * \log(1 + \exp(-t)*(\exp(- \rho) - 1)).}
   The copula is achieved by applying these functions to uniform random variables
   created from the univariate distibutions. Letting \eqn{u_i, i = 1, \dots, K}
   represent the realisations of the uniform random variables, Frank's copula can be
   calculated using
      \deqn{C_\rho(u_1, \dots, u_K) = \varphi^{[-1]}_\rho(
         \sum_{i=1}^K \varphi_\rho(u_i)).}
   The joint modelling of the univariate PoARX models form the multivariate
   PoARX model.
}
\references{
\insertRef{AgostoEtAl2016}{PoARX}

\insertRef{HallidayBoshnakov2018arxiv}{PoARX}

\insertRef{HallidayBoshnakov2018itise}{PoARX}

\insertRef{Joe2005}{PoARX}
}
