% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXMethods.R
\name{PoARX_methods}
\alias{PoARX_methods}
\alias{AIC.PoARX}
\alias{coef.PoARX}
\alias{df.residual.PoARX}
\alias{fitted.PoARX}
\alias{logLik.PoARX}
\alias{nobs.PoARX}
\alias{print.PoARX}
\alias{residuals.PoARX}
\alias{se.coef.PoARX}
\alias{summary.PoARX}
\alias{vcov.PoARX}
\title{Methods for PoARX objects}
\usage{
\method{AIC}{PoARX}(object, ...)

\method{coef}{PoARX}(object, ...)

\method{df.residual}{PoARX}(object, ...)

\method{fitted}{PoARX}(object, ...)

\method{logLik}{PoARX}(object, ...)

\method{nobs}{PoARX}(object, ...)

\method{print}{PoARX}(x, ...)

\method{residuals}{PoARX}(object, ...)

se.coef.PoARX(object, ...)

\method{summary}{PoARX}(object, ...)

\method{vcov}{PoARX}(object, ...)
}
\arguments{
\item{object, x}{A \code{"PoARX"} object from \code{PoARXFit}.}

\item{...}{Further arguments.}
}
\description{
Standard methods for PoARX objects.
}
\details{
Objects from class \code{"PoARX"} represent a fitted PoARX model, independent or
   dependent, as fitted by \code{indPoARXFit} or \code{depPoARXFit}, respectively.
}
\examples{
#### Examples go here!

}
\keyword{methods}
