% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXConstraints.R
\name{PoARXConstraints}
\alias{PoARXConstraints}
\title{Find the constraints for PoARX model(s) of the same structure}
\usage{
PoARXConstraints(M, r, rho = FALSE, intercept = TRUE, ylags = NULL,
  mulags = NULL, ...)
}
\arguments{
\item{M}{the number of sets of PoARX coefficients required.}

\item{r}{the number of exogenous covariates.}

\item{rho}{a logical indicator for whether a column is needed for a dependence
parameter.}

\item{intercept}{A logical indicator for whether an intercept should be
included in the model.}

\item{ylags}{A vector of time lags in the filter equation relating to the
observations.}

\item{mulags}{A vector of time lags in the filter equation relating to the
intensities.}

\item{...}{currently not used.}
}
\value{
A list containing the following:
    \item{ui}{A constraint matrix.}
    \item{ci}{A constraint vector.}
}
\description{
There are various constraints placed on the parameters \eqn{\omega},
   {\eqn{\alpha_i}}, {\eqn{\beta_i}}, and \eqn{\eta} in a PoARX model.
   This function creates matrices which can be used to implement such
   constraints in \code{constrOptim}.
}
\details{
The conditions for the parameters in a PoARX model are as follows. Firstly,
   all parameters must be non-negative to ensure that the Poisson parameter
   \eqn{\lambda} is non-negative. Secondly, the autoregressive parameters
   {\eqn{\alpha_i}} and {\eqn{\beta_i}} must satisfy
      \deqn{ \sum (\alpha_i + \beta_i) < 1 }
   in order to abide by the stationarity and ergodicity conditions.
}
