% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXFilter.R
\name{PoARXFilter}
\alias{PoARXFilter}
\title{Filter mean values for time series following a PoARX model}
\usage{
PoARXFilter(theta, obs, exogenous = NULL, intercept = TRUE,
  ylags = NULL, mulags = NULL, yinit = NULL, muinit = NULL, ...)
}
\arguments{
\item{theta}{A vector of parameter values.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{exogenous}{A 3-dimensional array containing the values of the exogenous
covariates. Each new value in dimension 1 represents another observation
of the time series. Each new vector in dimension 2 corresponds to a
new covariate and each matrix of the third dimension represents a
separate time series.}

\item{intercept}{A logical indicator for the presence of an intercept in the model.}

\item{ylags}{A vector of time lags in the filter equation relating to the
observations.}

\item{mulags}{A vector of time lags in the filter equation relating to the
intensities.}

\item{yinit}{A matrix detailing the initialisations of the observations.}

\item{muinit}{A matrix detailing the initialisations of the means.}

\item{...}{currently not used.}
}
\value{
A list containing the following:
    \item{Y}{A list of three elements: 'All', 'Initial', and 'Observed'.
       Each contains a matrix with the relevant y-values.}
    \item{Lambda}{A list of three elements: 'All', 'Initial', and 'Calculated'.
       Each contains a matrix with the relevant mean values.}
}
\description{
Using the observations and initialisation values, the PoARX filter
equation estimates the unobserved mean values.

The PoARX filter equation is
   \deqn{\lambda_t = \omega + \sum \alpha_l Y_{t-j_l} + \sum \beta_l
      \lambda_{t-j_l} + \eta x_{t-1}}
   where \eqn{\lambda} represents the mean of the process, \eqn{Y} the
   observations and \eqn{x} the covariates. \eqn{\omega}, {\eqn{\alpha_i}},
   {\eqn{\beta_i}} and \eqn{\eta} are all positive parameters that must be
   supplied.
}
