% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoARXMethods.R
\name{predict.PoARX}
\alias{predict.PoARX}
\title{Predict method for PoARX objects}
\usage{
\method{predict}{PoARX}(object, newdata = NULL, type = c("response",
  "probability"), ...)
}
\arguments{
\item{object}{A \code{"PoARX"} object from \code{PoARXFit}.}

\item{newdata}{An optional data.frame for the new covariates and observations.
The names of the data.frame should match the original variable names. If some new
observations have not been obtained, put an NA value where the missing value is
and the observation will be estimated by \code{PoARXPredict}.}

\item{type}{A character argument indicating the type of response. If \code{"response"},
return the mean parameter for each time series. If \code{"probability"}, give the
probability of the response.}

\item{...}{Further arguments.}
}
\value{
If \code{type = "response"} a data.frame containing
   \eqn{\lambda} values for each time series, along with optional confidence intervals.

   If \code{type = "probability"} and the respective observations are provided, then
   a matrix of probabilities is returned. If no observations are given, there
   are too many possible values, so the \eqn{\lambda}{lambda} values are returned instead,
   as if \code{type = "response"}.
}
\description{
The \code{predict} function can be used to predict the next h steps
   given new data or return fitted value. If \code{newdata} is \code{NULL}, then
   fitted values will be returned if \code{"response"} is chosen, otherwise
   the probability of each observed data point. Missing values can be represented by
   \code{NA} and are dealt with by \code{PoARXPredict}. Note that as h increases,
   the variances of the predictions will increase. Confidence intervals can be
   returned if desired, using simulation.
}
\examples{
#### Examples go here!

}
