% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaFunctions.R
\name{rhoFCLogLikelihood}
\alias{rhoFCLogLikelihood}
\alias{rhoFCScore}
\title{Profile log-likelihood w.r.t. the dependence parameter rho}
\usage{
rhoFCLogLikelihood(rho, obs, means, zeroFix = 1e-35, ...)

rhoFCScore(rho, obs, means, zeroFix = 1e-35)
}
\arguments{
\item{rho}{A numeric value specifying the dependence parameter used in
\code{pFranksCopula}.}

\item{obs}{A matrix of discrete-valued time series observations. Each
column of the matrix represents a separate time series.}

\item{means}{A matrix of calculated means, from \code{PoARXFilter}. It should
have the same dimensions as \code{obs}.}

\item{zeroFix}{A numeric value to transform any 0-probability values to.}

\item{...}{currently not used.}
}
\value{
for \code{rhoFCLogLikelihood}, a numeric value representing the
    log-likelihood of the data for dependence parameter rho.

for \code{rhoFCScore}, a numeric value representing the score.
}
\description{
Profile log-likelihood w.r.t. the dependence parameter rho
}
\details{
During the IMF process the copula parameters are estimated
    separately from the marginal parameters.

   \code{rhoFCLogLikelihood} computes the full log-likelihood of Poisson
   models coupled using Franks copula with respect to the variation of rho
   alone. \strong{(TODO: This sounds bizzare)}

\code{rhoFCScore} computes the profile score function for the
    dependence parameter rho, i.e. the derivative of the log-likelihood with
    respect to rho.
}
